% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch BiasCorrector}
\usage{
launch_app(
  port = 3838,
  plotdir = "plots",
  csvdir = "csv",
  logfilename = "biascorrector.log",
  maxfilesize = 100
)
}
\arguments{
\item{port}{The port, BiasCorrector is running on (default: 3838)}

\item{plotdir}{A character string. Defaults to 'plots'. This directory
is being created inside tempdir.}

\item{csvdir}{A character string. Defaults to 'csv'. This directory is
being created inside tempdir.}

\item{logfilename}{A character string. The name of the logfile
(default = biascorrector.log).}

\item{maxfilesize}{A positive integer. The maximum file size allowed
for upload.}
}
\value{
BiasCorrector shiny application
}
\description{
Launch BiasCorrector
}
