% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launchApp}
\alias{launchApp}
\title{Launch BiasCorrector}
\usage{
launchApp(port = 3838, maxfilesize = 100,
  logfilename = "biascorrector.log", plotdir = "plots",
  csvdir = "csv")
}
\arguments{
\item{port}{The port, BiasCorrector is running on (default: 3838)}

\item{maxfilesize}{A positive integer. The maximum file size allowed for upload.}

\item{logfilename}{A character string. The name of the logfile (default = biascorrector.log).}

\item{plotdir}{A character string. Defaults to 'plots'. This directory is being created inside tempdir.}

\item{csvdir}{A character string. Defaults to 'csv'. This directory is being created inside tempdir.}
}
\value{
BiasCorrector shiny application
}
\description{
Launch BiasCorrector
}
