% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCorrectedPlots.R
\name{module_correctedplots_ui}
\alias{module_correctedplots_ui}
\title{module_correctedplots_ui}
\usage{
module_correctedplots_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_correctedplots_ui
}
