% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCorrectedStatistics.R
\name{module_correctedstatistics_ui}
\alias{module_correctedstatistics_ui}
\title{module_correctedstatistics_ui}
\usage{
module_correctedstatistics_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_correctedstatistics_ui
}
