% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExperimentalFile.R
\name{module_experimentalfile_server}
\alias{module_experimentalfile_server}
\title{module_experimentalfile_server}
\usage{
module_experimentalfile_server(input, output, session, rv)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny session object}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}
}
\description{
module_experimentalfile_server
}
