% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExperimentalFile.R
\name{module_experimentalfile_ui}
\alias{module_experimentalfile_ui}
\title{module_experimentalfile_ui}
\usage{
module_experimentalfile_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_experimentalfile_ui
}
