% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFileupload.R
\name{module_fileupload_ui}
\alias{module_fileupload_ui}
\title{module_fileupload_ui}
\usage{
module_fileupload_ui(id, ...)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}

\item{...}{Further arguments, such as `maximumfilesize`}
}
\description{
module_fileupload_ui
}
