% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleModelSelection.R
\name{module_modelselection_server}
\alias{module_modelselection_server}
\title{module_modelselection_server}
\usage{
module_modelselection_server(input, output, session, rv, input_re)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny session object}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}

\item{input_re}{The Shiny server input object, wrapped into a reactive
expression: input_re = reactive({input})}
}
\value{
The function returns a shiny server module.
}
\description{
module_modelselection_server
}
\examples{
if (interactive()) {
rv <- list()
logfilename <- paste0(tempdir(), "/log.txt")
shiny::callModule(
  module_modelselection_server,
  "moduleModelSelection",
  rv = rv,
  logfilename = logfilename
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
