% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleSettings.R
\name{module_settings_ui}
\alias{module_settings_ui}
\title{module_settings_ui}
\usage{
module_settings_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
module_settings_ui
}
