% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCalibrationFile.R
\name{moduleCalibrationFileUI}
\alias{moduleCalibrationFileUI}
\title{moduleCalibrationFileUI}
\usage{
moduleCalibrationFileUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleCalibrationFileUI
}
