% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCorrectedPlots.R
\name{moduleCorrectedPlotsUI}
\alias{moduleCorrectedPlotsUI}
\title{moduleCorrectedPlotsUI}
\usage{
moduleCorrectedPlotsUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleCorrectedPlotsUI
}
