% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCorrectedStatistics.R
\name{moduleCorrectedStatisticsUI}
\alias{moduleCorrectedStatisticsUI}
\title{moduleCorrectedStatisticsUI}
\usage{
moduleCorrectedStatisticsUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleCorrectedStatisticsUI
}
