% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExperimentalFile.R
\name{moduleExperimentalFileServer}
\alias{moduleExperimentalFileServer}
\title{moduleExperimentalFileServer}
\usage{
moduleExperimentalFileServer(input, output, session, rv)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny session object}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}
}
\description{
moduleExperimentalFileServer
}
