% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExperimentalFile.R
\name{moduleExperimentalFileUI}
\alias{moduleExperimentalFileUI}
\title{moduleExperimentalFileUI}
\usage{
moduleExperimentalFileUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleExperimentalFileUI
}
