% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFileupload.R
\name{moduleFileuploadUI}
\alias{moduleFileuploadUI}
\title{moduleFileuploadUI}
\usage{
moduleFileuploadUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleFileuploadUI
}
