% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleModelSelection.R
\name{moduleModelSelectionUI}
\alias{moduleModelSelectionUI}
\title{moduleModelSelectionUI}
\usage{
moduleModelSelectionUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleModelSelectionUI
}
