% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePlotting.R
\name{modulePlottingUI}
\alias{modulePlottingUI}
\title{modulePlottingUI}
\usage{
modulePlottingUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
modulePlottingUI
}
