% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleResults.R
\name{moduleResultsUI}
\alias{moduleResultsUI}
\title{moduleResultsUI}
\usage{
moduleResultsUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleResultsUI
}
