% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleSettings.R
\name{moduleSettingsUI}
\alias{moduleSettingsUI}
\title{moduleSettingsUI}
\usage{
moduleSettingsUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleSettingsUI
}
