% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleStatistics.R
\name{moduleStatisticsUI}
\alias{moduleStatisticsUI}
\title{moduleStatisticsUI}
\usage{
moduleStatisticsUI(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\description{
moduleStatisticsUI
}
