% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launchApp}
\alias{launchApp}
\title{Launch BiasCorrector}
\usage{
launchApp(port = 3838)
}
\arguments{
\item{port}{The port, BiasCorrector is running on (default: 3838)}
}
\value{
BiasCorrector shiny application
}
\description{
Launch BiasCorrector
}
