% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCorrectedPlots.R
\name{module_correctedplots_ui}
\alias{module_correctedplots_ui}
\title{module_correctedplots_ui}
\usage{
module_correctedplots_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_correctedplots_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "correctedplots",
    module_correctedplots_ui(
      "moduleCorrectedPlots"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
