% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCorrectedStatistics.R
\name{module_correctedstatistics_ui}
\alias{module_correctedstatistics_ui}
\title{module_correctedstatistics_ui}
\usage{
module_correctedstatistics_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_correctedstatistics_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "correctedstats",
    module_correctedstatistics_ui(
      "moduleCorrectedStats"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
