% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExperimentalFile.R
\name{module_experimentalfile_server}
\alias{module_experimentalfile_server}
\title{module_experimentalfile_server}
\usage{
module_experimentalfile_server(input, output, session, rv, ...)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{Shiny session object}

\item{rv}{The global 'reactiveValues()' object, defined in server.R}

\item{...}{Further arguments, such as `logfilename`, `csvdir` and `plotdir`}
}
\value{
The function returns a shiny server module.
}
\description{
module_experimentalfile_server
}
\examples{
if (interactive()) {
rv <- list()
logfilename <- paste0(tempdir(), "/log.txt")
shiny::callModule(
  module_experimentalfile_server,
  "moduleExperimentalFile",
  rv = rv,
  logfilename = logfilename
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
