% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExperimentalFile.R
\name{module_experimentalfile_ui}
\alias{module_experimentalfile_ui}
\title{module_experimentalfile_ui}
\usage{
module_experimentalfile_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_experimentalfile_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "experimental",
    module_experimentalfile_ui(
      "moduleExperimentalFile"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
