% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFileupload.R
\name{module_fileupload_ui}
\alias{module_fileupload_ui}
\title{module_fileupload_ui}
\usage{
module_fileupload_ui(id, ...)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}

\item{...}{Further arguments, such as `maxfilesize`}
}
\value{
The function returns a shiny ui module.
}
\description{
module_fileupload_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "fileupload",
    module_fileupload_ui(
      "moduleFileUpload",
      maxfilesize = maxfilesize
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
