% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePlotting.R
\name{module_plotting_ui}
\alias{module_plotting_ui}
\title{module_plotting_ui}
\usage{
module_plotting_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_plotting_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "plotting",
    module_plotting_ui(
      "modulePlotting"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
