% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleResults.R
\name{module_results_ui}
\alias{module_results_ui}
\title{module_results_ui}
\usage{
module_results_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_results_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "results",
    module_results_ui(
      "moduleResults"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
