% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleSettings.R
\name{module_settings_ui}
\alias{module_settings_ui}
\title{module_settings_ui}
\usage{
module_settings_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_settings_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "settings",
    module_settings_ui(
      "moduleSettings"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
