% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleStatistics.R
\name{module_statistics_ui}
\alias{module_statistics_ui}
\title{module_statistics_ui}
\usage{
module_statistics_ui(id)
}
\arguments{
\item{id}{A character. The identifier of the shiny object}
}
\value{
The function returns a shiny ui module.
}
\description{
module_statistics_ui
}
\examples{
if (interactive()) {
shinydashboard::tabItems(
  shinydashboard::tabItem(
    tabName = "statistics",
    module_statistics_ui(
      "moduleStatistics"
    )
  )
)
}

}
\seealso{
\url{https://shiny.rstudio.com/articles/modules.html}
}
